+++ src/OpenTelemetry/Resources/Resource.cs
@@ -109,7 +109,12 @@ namespace OpenTelemetry.Resources
         {
             if (value != null)
             {
-                if (value is string || value is bool || value is long || value is double)
+                if (value is string || value is bool || value is double || value is long)
+                {
+                    return value;
+                }
+
+                if (value is string[] || value is bool[] || value is double[] || value is long[])
                 {
                     return value;
                 }
@@ -122,6 +127,32 @@ namespace OpenTelemetry.Resources
                 if (value is float)
                 {
                     return System.Convert.ToDouble(value, System.Globalization.CultureInfo.InvariantCulture);
+                }
+
+                if (value is int[] || value is short[])
+                {
+                    long[] convertedArr = new long[((System.Array)value).Length];
+                    int i = 0;
+                    foreach (var val in (System.Array)value)
+                    {
+                        convertedArr[i] = System.Convert.ToInt64(val);
+                        i++;
+                    }
+
+                    return convertedArr;
+                }
+
+                if (value is float[])
+                {
+                    double[] convertedArr = new double[((float[])value).Length];
+                    int i = 0;
+                    foreach (float val in (float[])value)
+                    {
+                        convertedArr[i] = System.Convert.ToDouble(val, System.Globalization.CultureInfo.InvariantCulture);
+                        i++;
+                    }
+
+                    return convertedArr;
                 }
 
                 throw new System.ArgumentException("Attribute value type is not an accepted primitive", keyName);